repeater

+++ dominio/Concesionaria.cs
@@ -13,10 +13,10 @@ namespace dominio
         public string Descripcion { get; set; }
         public string ImagenUrl { get; set; }
         public string MapsUrl { get; set; }
-        public int SemanaAbre { get; set; }
-        public int SemanaCierra { get; set; }
-        public int SabadoAbre { get; set; }
-        public int SabadoCierra { get; set; }
+        public decimal SemanaAbre { get; set; }
+        public decimal SemanaCierra { get; set; }
+        public decimal SabadoAbre { get; set; }
+        public decimal SabadoCierra { get; set; }
         public string Localidad { get; set; }
         public string Provincia { get; set; }
     }
